requirements.txt
# Function dependencies, for example:
# package>=version
lxml
sendgrid
flask
requests
google-cloud-logging